BE 2주차 세션 코드(로그인/회원가입/권한관리/미디어)

+++ main/views.py
@@ -13,13 +13,19 @@ def blogpage(request):
     return render(request, 'main/blogpage.html', {'blogs': blogs})
 
 def new_blog(request):
+    if not request.user.is_authenticated:
+        return redirect('accounts:login')
+
     return render(request, 'main/new_blog.html')
 
 def create(request):
+    if not request.user.is_authenticated:
+        return redirect('accounts:login')
+
     new_blog = Blog()
 
     new_blog.title = request.POST['title']
-    new_blog.writer = request.POST['writer']
+    new_blog.writer = request.user.username
     new_blog.pub_date = request.POST['pub_date']
     new_blog.content = request.POST['content']
 
@@ -33,12 +39,26 @@ def detail(request, blog_id):
 
 def edit(request, blog_id):
     edit_blog = get_object_or_404(Blog, pk=blog_id)
+
+    if not request.user.is_authenticated:
+        return redirect('accounts:login')
+
+    if edit_blog.writer != request.user.username:
+        return redirect('main:detail', edit_blog.id)
+
     return render(request, 'main/edit.html', {"blog": edit_blog})
 
 def update(request, blog_id):
     update_blog = get_object_or_404(Blog, pk=blog_id)
+
+    if not request.user.is_authenticated:
+        return redirect('accounts:login')
+
+    if update_blog.writer != request.user.username:
+        return redirect('main:detail', update_blog.id)
+
     update_blog.title = request.POST['title']
-    update_blog.writer = request.POST['writer']
+    update_blog.writer = request.user.username
     update_blog.pub_date = request.POST['pub_date']
     update_blog.content = request.POST['content']
     update_blog.save()
@@ -47,6 +67,13 @@ def update(request, blog_id):
 
 def delete(request, blog_id):
     delete_blog = get_object_or_404(Blog, pk=blog_id)
+
+    if not request.user.is_authenticated:
+        return redirect('accounts:login')
+
+    if delete_blog.writer != request.user.username:
+        return redirect('main:detail', delete_blog.id)
+
     delete_blog.delete()
 
     return redirect('main:blogpage')
